feat(img_cache): add support for alternate img cache algorithm
