+++ src/draw/lv_img_cache_builtin.h
@@ -1,0 +1,39 @@
+/**
+ * @file lv_img_cache_builtin.h
+ *
+ */
+
+#ifndef LV_IMG_CACHE_BUILTIN_H
+#define LV_IMG_CACHE_BUILTIN_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/*********************
+ *      INCLUDES
+ *********************/
+
+/*********************
+ *      DEFINES
+ *********************/
+
+/**********************
+ *      TYPEDEFS
+ **********************/
+
+/**********************
+ * GLOBAL PROTOTYPES
+ **********************/
+
+void lv_img_cache_builtin_init(void);
+
+/**********************
+ *      MACROS
+ **********************/
+
+#ifdef __cplusplus
+} /*extern "C"*/
+#endif
+
+#endif /*LV_IMG_CACHE_BUILTIN_H*/
